Try Float.NaN

+++ python/src/test/scala/latis/AnomalyDataSpec.scala
@@ -58,7 +58,7 @@ class AnomalyDataSpec extends FlatSpec {
       case Sample(DomainData(Number(t)), RangeData(Real(f), Real(a))) =>
         t should be (1)
         f should be (0.841470985)
-        a should be (Double.NaN)
+        a should be (Float.NaN)
     }
   }
